refactor: centralize auditable properties
